+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPlugin.kt
@@ -129,8 +129,8 @@ open class IntelliJPlugin : Plugin<Project> {
     }
 
     private fun checkGradleVersion(project: Project) {
-        if (Version.parse(project.gradle.gradleVersion) < Version.parse("6.7")) {
-            throw PluginInstantiationException("${IntelliJPluginConstants.NAME} requires Gradle 6.7 and higher")
+        if (Version.parse(project.gradle.gradleVersion) < Version.parse("6.7.1")) {
+            throw PluginInstantiationException("${IntelliJPluginConstants.NAME} requires Gradle 6.7.1 and higher")
         }
     }