+++ CHANGES.md
@@ -7,6 +7,9 @@
 - Set `ResolutionStrategy.SortOrder.DEPENDENCY_FIRST` for `compileClasspath` and `testCompileClasspath` configurations [#656](../../issues/656)
 - Added `useDependencyFirstResolutionStrategy` feature flag. See [Feature Flags](https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html#build-features).
 
+### Changed
+- Set minimal supported Gradle version from `6.7` to `6.7.1`
+
 ### Fixed
 - Fixed broken instrumentation when custom sources directory is set [#1004](../../issues/1004)
 - Fixed `java.nio.file.FileAlreadyExistsException: instrumentCode/classpath.index` [#998](../../issues/998)